Move lazy evaluation code to its own module
