+++ test/unit/test_laziness.py
@@ -39,6 +39,8 @@ import pytest
 import numpy
 
 from pyop2 import op2
+from pyop2 import lazy
+
 
 nelems = 42
 
@@ -46,34 +48,34 @@ nelems = 42
 class TestExecutionTrace:
 
     def test_read_write_dependency(self, backend, skip_greedy):
-        a = op2.base.LazyPass(set([1]), set())
-        b = op2.base.LazyPass(set(), set([1]))
+        a = lazy.LazyPass(set([1]), set())
+        b = lazy.LazyPass(set(), set([1]))
 
         assert b.depends_on(a)
 
     def test_write_read_dependency(self, backend, skip_greedy):
-        a = op2.base.LazyPass(set(), set([1]))
-        b = op2.base.LazyPass(set([1]), set())
+        a = lazy.LazyPass(set(), set([1]))
+        b = lazy.LazyPass(set([1]), set())
 
         assert b.depends_on(a)
 
     def test_maintain_write_order(self, backend, skip_greedy):
-        a = op2.base.LazyPass(set(), set([1]))
-        b = op2.base.LazyPass(set(), set([1]))
+        a = lazy.LazyPass(set(), set([1]))
+        b = lazy.LazyPass(set(), set([1]))
 
         assert b.depends_on(a)
 
     def test_empty_trace(self, backend, skip_greedy):
         """Check trace initial state."""
-        trace = op2.base.ExecutionTrace()
+        trace = lazy.ExecutionTrace()
         assert trace.in_queue(trace.top)
         assert trace.in_queue(trace.bot)
         assert trace.bot in trace.children(trace.top)
 
     def test_enqueue(self, backend, skip_greedy):
-        trace = op2.base.ExecutionTrace()
+        trace = lazy.ExecutionTrace()
 
-        c = op2.base.LazyPass(set(), set())
+        c = lazy.LazyPass(set(), set())
         trace._append(c)
 
         assert trace.in_queue(c)
@@ -81,10 +83,10 @@ class TestExecutionTrace:
         assert c in trace.parents(trace.bot)
 
     def test_siblings(self, backend, skip_greedy):
-        trace = op2.base.ExecutionTrace()
+        trace = lazy.ExecutionTrace()
 
-        a = op2.base.LazyPass(set([1]), set([1]))
-        b = op2.base.LazyPass(set([2]), set([2]))
+        a = lazy.LazyPass(set([1]), set([1]))
+        b = lazy.LazyPass(set([2]), set([2]))
 
         trace._append(a)
         trace._append(b)
@@ -94,11 +96,11 @@ class TestExecutionTrace:
         assert len(trace.children(trace.top)) == 2
 
     def test_children(self, backend, skip_greedy):
-        trace = op2.base.ExecutionTrace()
+        trace = lazy.ExecutionTrace()
 
-        a = op2.base.LazyPass(set([1]), set([1]))
-        b = op2.base.LazyPass(set([1]), set([2]))
-        c = op2.base.LazyPass(set([1]), set([3]))
+        a = lazy.LazyPass(set([1]), set([1]))
+        b = lazy.LazyPass(set([1]), set([2]))
+        c = lazy.LazyPass(set([1]), set([3]))
 
         trace._append(a)
         trace._append(b)
@@ -137,7 +139,7 @@ count(unsigned int* x)
 
     def test_ro_accessor(self, backend, skip_greedy, iterset):
         """Read-only access to a Dat should force computation that writes to it."""
-        op2.base._trace.clear()
+        lazy._trace.clear()
         d = op2.Dat(iterset, numpy.zeros(iterset.total_size), dtype=numpy.float64)
         k = op2.Kernel('void k(double *x) { *x = 1.0; }', 'k')
         op2.par_loop(k, iterset, d(op2.WRITE))
@@ -146,7 +148,7 @@ count(unsigned int* x)
     def test_rw_accessor(self, backend, skip_greedy, iterset):
         """Read-write access to a Dat should force computation that writes to it,
         and any pending computations that read from it."""
-        op2.base._trace.clear()
+        lazy._trace.clear()
         d = op2.Dat(iterset, numpy.zeros(iterset.total_size), dtype=numpy.float64)
         d2 = op2.Dat(iterset, numpy.empty(iterset.total_size), dtype=numpy.float64)
         k = op2.Kernel('void k(double *x) { *x = 1.0; }', 'k')
